silence warnings about convert double to int
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@3084 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ testing/testing_zheevdx_2stage.cpp
@@ -127,7 +127,7 @@ int main( int argc, char** argv)
             magma_int_t iu = 0;
             if (range == 'I'){
                 il = 1;
-                iu = f*N;
+                iu = (int) f*N;
             }
 
             if(opts.warmup){
